backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
Pillow
channels
daphne
channels-redis